% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyVariants.R
\name{classifyVariants}
\alias{classifyVariants}
\title{Classify Candidate Variants Based on Haplotype Analysis}
\usage{
classifyVariants(
  candidate_variant_granges,
  phasing_orientation = c("left", "right"),
  haplotype_granges,
  haplotype_boundary_coordinate_granges,
  boundary_cutoff,
  IBD_distance_cutoff,
  non_IBD_distance_cutoff,
  PS_width_cutoff,
  QC_fail_variant_granges
)
}
\arguments{
\item{candidate_variant_granges}{A \code{GRanges} object containing the candidate variants to be classified.}

\item{phasing_orientation}{A character value indicating the phasing orientation ("left" or "right").}

\item{haplotype_granges}{A \code{GRanges} object containing haplotype information of the sequenced parent.}

\item{haplotype_boundary_coordinate_granges}{A \code{GRanges} object containing the haplotype block boundaries.}

\item{boundary_cutoff}{A numeric value indicating the distance from the boundary of a haplotype block
within which candidate variants are considered.}

\item{IBD_distance_cutoff}{A numeric value specifying the minimum Hamming distance required for inferring
the non-IBD status of a haplotype pair.}

\item{non_IBD_distance_cutoff}{A numeric value specifying the maximum Hamming distance allowed to infer
the IBD status of a haplotype pair.}

\item{PS_width_cutoff}{A numeric value specifying the minimum width for phasing sets to be included in the analysis.}

\item{QC_fail_variant_granges}{A \code{GRanges} object containing the candidate variants that have failed QC thus far (e.g. due to low sequencing depth or GQ).}
}
\value{
A \code{GRanges} object containing classified candidate variants. The classified variants include
the following columns:
\itemize{
\item \code{duoNovo_classification}: Classification status ("de novo", "on other parent haplotype", or "uncertain").
\item \code{hamming_distance_other_parent_hap}: The minimum Hamming distance for the other parent's haplotype.
\item \code{supporting_counts_het_hom}, \code{supporting_counts_het_het}, \code{supporting_counts_hom_het}:
Counts representing variant pairs used to calculate Hamming distance.
}
}
\description{
This function classifies candidate variants as de novo or inherited based on the comparison
of phased haplotype blocks between the proband and the sequenced parent.
}
\examples{
# classifyVariants(candidate_variant_granges = candidate_variants,
#                phasing_orientation = "left",
#                 haplotype_granges = haplotypes,
#                haplotype_boundary_coordinate_granges = haplotype_boundaries,
#                  boundary_cutoff = 2000,
#                  distance_cutoff = 40)

}
