% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoNovo.R
\name{duoNovo}
\alias{duoNovo}
\title{Identify de novo variants from duo (proband - single parent) long read sequencing}
\usage{
duoNovo(
  LRS_phased_vcf_file_path,
  depth_cutoff = 20,
  GQ_cutoff = 30,
  proband_phasing,
  proband_column_identifier,
  PS_width_cutoff = 10000,
  boundary_cutoff = 2000,
  distance_cutoff = 40,
  candidate_variants_concordant_with_SRS = TRUE,
  SRS_vcf_file_path
)
}
\arguments{
\item{LRS_phased_vcf_file_path}{File path to the vcf containing phased variant calls from long-read sequencing of the duo.}

\item{depth_cutoff}{A numeric value specifying the minimum sequencing depth for variants to be included in the analysis. The same cutoff applies to both proband and parent variants, for both LRS and SRS (if used).}

\item{GQ_cutoff}{A numeric value specifying the minimum GQ (genotype quality) for variants to be included in the analysis. The same cutoff applies to both proband and parent variants, for both LRS and SRS (if used).}

\item{proband_phasing}{A character string indicating the phasing of the proband, either "1|0" or "0|1".}

\item{proband_column_identifier}{A character corresponding to an identifier for the proband column in the metadata matrices. Should be the same for both the LRS vcf and (if used) the SRS vcf.}

\item{PS_width_cutoff}{A numeric value specifying the minimum width for phasing sets to be included in the analysis.}

\item{boundary_cutoff}{A numeric value indicating the minimum distance from a haplotype block boundary (either start or end coordinate) for candidate variants to be analyzed.}

\item{distance_cutoff}{A numeric value specifying the minimum hamming distance cutoff to determine that a proband-parent haplotype block are not identical by descent.}

\item{candidate_variants_concordant_with_SRS}{Logical value specifying if candidate variants should be concordant with short-read sequencing (default is \code{TRUE} or \code{FALSE}).}

\item{SRS_vcf_file_path}{File path to the vcf containing variant calls from short-read sequencing of the duo.}
}
\value{
A \code{GRangesList} containing three elements: \code{de_novo}, \code{not_de_novo}, and \code{uncertain}.
}
\description{
The \code{duoNovo} function uses phased variant calls from duo long-read sequencing to classify candidate variants (heterozygous in the proband; absent in the parent).
Variants are classifed as de novo, present on the haplotype inherited from the non-sequenced parent, or uncertain variants.
}
\details{
The function ultimately works by detecting identical by descent haplotype blocks, to determine whether each candidate variant of interest is de novo, using the genotype of only one parent. If requested, concordance with short-read sequencing can be checked.
}
\examples{
# duoNovo(LRS_phased_vcf_file_path = my_LRS_file_path,
#          proband_phasing = "1|0", proband_column_identifier = "-0$",
#          PS_width_cutoff = 10000, boundary_cutoff = 10000, distance_cutoff = 2000,
#          candidate_variants_concordant_with_SRS = TRUE,
#          SRS_vcf_file_path = my_SRS_file_path)
}
