% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaplotypeBlockCoordinates.R
\name{getHaplotypeBlockCoordinates}
\alias{getHaplotypeBlockCoordinates}
\title{Get haplotype block boundaries}
\usage{
getHaplotypeBlockCoordinates(haplotype_granges)
}
\arguments{
\item{haplotype_granges}{A \code{GRanges} object containing haplotype information with \code{PS1} and \code{PS2} columns representing the phasing sets of the proband and parent, respectively.}
}
\value{
A \code{GRanges} object with combined coordinates of overlapping phasing sets, including \code{PS1} and \code{PS2} identifiers.
}
\description{
The \code{getHaplotypeBlockCoordinates} function takes as input a \code{GRanges} object containing haplotype information, and identifies genomic segments where proband variants are assigned to the same phasing set, and parental variants are assigned to the same phasing set.
The function combines phasing set boundaries (\code{PS1} and \code{PS2}) from the proband and parent phasing (using HiPhase), creating a unified \code{GRanges} object.
}
\details{
The function splits haplotype ranges by \code{PS1} and \code{PS2} identifiers, calculates start and end coordinates for each group,
and then finds overlapping phasing sets between \code{PS1} and \code{PS2}. The resulting overlapping coordinates are stored in a new \code{GRanges} object.
}
\examples{
# Assuming `haplotype_granges` is defined and contains `PS1` and `PS2` columns:
# combined_coordinates <- getPhasingSetCoordinates(haplotype_granges)
}
