% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaplotypes.R
\name{getHaplotypes}
\alias{getHaplotypes}
\title{Extract haplotypes from phased variant calls}
\usage{
getHaplotypes(phased_variant_granges_duo)
}
\arguments{
\item{phased_variant_granges_duo}{A \code{GRanges} object containing phased variant calls for both the proband and parent. Generated with the VariantAnnotation package from a phased vcf as input. We also recommend filtering based on sequencing depth and genotype quality prior to extracting haplotypes.}
}
\value{
A \code{GRanges} object with additional columns (\code{hap11}, \code{hap21}, \code{hap12}, \code{hap22}) indicating phased alleles for both individuals.
}
\description{
The \code{getHaplotypes} function extracts haplotype information from a \code{GRanges} object containing phased variant calls for both proband and parent.
Each of the two haplotypes of an individual is represented in a separate column.
}
\details{
The function identifies the haplotype data for the proband and parent from the \code{phasing1} and \code{phasing2} columns.
The phased alleles are categorized for each variant using regular expressions to determine the phase (e.g., \code{0|1}, \code{1|0}, etc.).
}
\examples{
# Assuming `phased_variant_granges_duo` is defined and contains the appropriate phasing columns:
# phased_haplotypes <- getHaplotypes(phased_variant_granges_duo)
}
